backend/requirements.txt

fastapi
uvicorn
python-multipart
pytesseract
pillow
pdf2image
spacy
mysql-connector-python
sqlalchemy
bcrypt
reportlab
python-docx
cryptography
opencv-python
pytest
tomli
python-dotenv
openai
chromadb
numpy
pymupdf
